continue removing store
