+++ Projects/Claims/Sources/Views/SubmitClaim/SubViews/SupportView.swift
@@ -6,6 +6,7 @@ import hCoreUI
 
 struct SupportView: View {
     @PresentableStore var store: SubmitClaimStore
+    @EnvironmentObject var claimsNavigationVm: ClaimsNavigationViewModel
     let openChat: () -> Void
 
     var body: some View {
@@ -18,7 +19,7 @@ struct SupportView: View {
                     .multilineTextAlignment(.center)
             }
             hButton.MediumButton(type: .primary) {
-                store.send(.dismissNewClaimFlow)
+                claimsNavigationVm.router.dismiss()
                 DispatchQueue.main.asyncAfter(deadline: .now() + 2) {
                     openChat()
                 }
