+++ Projects/Claims/Sources/Views/SubmitClaim/SubmitClaimContactScreen.swift
@@ -27,7 +27,6 @@ public struct SubmitClaimContactScreen: View, KeyboardReadable {
                             placeholder: L10n.phoneNumberRowTitle,
                             error: $vm.phoneNumberError
                         )
-                        .disableOn(SubmitClaimStore.self, [.postPhoneNumber])
                         hButton.LargeButton(type: .primary) {
                             if vm.keyboardEnabled {
                                 withAnimation {
@@ -49,7 +48,6 @@ public struct SubmitClaimContactScreen: View, KeyboardReadable {
                         } content: {
                             hText(vm.keyboardEnabled ? L10n.generalSaveButton : L10n.generalContinueButton)
                         }
-                        .trackLoading(SubmitClaimStore.self, action: .postPhoneNumber)
                         .presentableStoreLensAnimation(.default)
                         .disabled(!(vm.enableContinueButton || vm.keyboardEnabled))
 
@@ -58,7 +56,6 @@ public struct SubmitClaimContactScreen: View, KeyboardReadable {
                 }
                 .sectionContainerStyle(.transparent)
             }
-            .claimErrorTrackerFor([.postContractSelect])
             .onReceive(keyboardPublisher) { keyboardHeight in
                 vm.keyboardEnabled = keyboardHeight != nil
             }
